+++ inst/benchmarks/_helpers.R
@@ -132,6 +132,8 @@ svd_surface_default_methods <- function(args) {
       "eigencore_irlba_lbd_one_sided",
       "eigencore_irlba_lbd_retained_native",
       "eigencore_irlba_lbd_retained_bpro",
+      "eigencore_irlba_lbd_retained_bpro_one_sided_guarded",
+      "eigencore_irlba_lbd_retained_bpro_block_guarded",
       "eigencore_irlba_lbd_normal_scout",
       "eigencore_golub_kahan_projected",
       "eigencore_implicit_normal_lanczos",
@@ -181,6 +183,8 @@ svd_internal_methods <- function() {
     "eigencore_irlba_lbd_one_sided",
     "eigencore_irlba_lbd_retained_native",
     "eigencore_irlba_lbd_retained_bpro",
+    "eigencore_irlba_lbd_retained_bpro_one_sided_guarded",
+    "eigencore_irlba_lbd_retained_bpro_block_guarded",
     "eigencore_irlba_lbd_normal_scout",
     "eigencore_golub_kahan_projected",
     "eigencore_implicit_normal_lanczos",
@@ -1497,6 +1501,18 @@ benchmark_svd_case <- function(A, rank, methods = NULL, iterations = 3L,
         result_restart_numeric(fit, "irlba_lbd_bpro_basis_orthogonality_loss"),
       irlba_lbd_bpro_escalation_recommended =
         result_restart_logical(fit, "irlba_lbd_bpro_escalation_recommended"),
+      irlba_lbd_reorth_mode =
+        result_restart_character(fit, "irlba_lbd_reorth_mode"),
+      irlba_lbd_one_sided_reorth_used =
+        result_restart_logical(fit, "irlba_lbd_one_sided_reorth_used"),
+      irlba_lbd_bpro_block_size =
+        result_restart_integer(fit, "irlba_lbd_bpro_block_size"),
+      irlba_lbd_bpro_exact_orthogonality_loss =
+        result_restart_numeric(fit, "irlba_lbd_bpro_exact_orthogonality_loss"),
+      irlba_lbd_bpro_exact_orthogonality_passed =
+        result_restart_logical(fit, "irlba_lbd_bpro_exact_orthogonality_passed"),
+      irlba_lbd_bpro_guard_fallback_reason =
+        result_restart_character(fit, "irlba_lbd_bpro_guard_fallback_reason"),
       irlba_lbd_retained_seed_strategy =
         result_restart_character(fit, "irlba_lbd_retained_seed_strategy"),
       irlba_lbd_retained_from_scout =